Separate the actions API from the core actions API.
