+++ rules.services.yml
@@ -2,6 +2,9 @@ services:
   plugin.manager.rules_expression:
     class: Drupal\rules\Engine\ExpressionPluginManager
     arguments: ['@container.namespaces', '@module_handler']
+  plugin.manager.rules_action:
+    class: Drupal\rules\Core\RulesActionManager
+    parent: default_plugin_manager
   plugin.manager.rules_data_processor:
     class: Drupal\rules\Context\DataProcessorManager
     arguments: ['@container.namespaces', '@module_handler']
