+++ src/Core/Annotation/RulesAction.php
@@ -1,0 +1,70 @@
+<?php
+
+/**
+ * @file
+ * Contains \Drupal\rules\Core\Annotation\RulesAction.
+ */
+
+namespace Drupal\rules\Core\Annotation;
+
+use Drupal\Component\Annotation\Plugin;
+
+/**
+ * Defines a Rules Action annotation object.
+ *
+ * Plugin Namespace: Plugin\RulesAction
+ *
+ * For a working example, see \Drupal\rules\Plugin\RulesAction\BanIP
+ *
+ * @see \Drupal\rules\Core\RulesActionInterface
+ * @see \Drupal\rules\Core\RulesActionManagerInterface
+ * @see \Drupal\rules\Core\RulesActionBase
+ * @see plugin_api
+ *
+ * @Annotation
+ */
+class RulesAction extends Plugin {
+
+  /**
+   * The plugin ID.
+   *
+   * @var string
+   */
+  public $id;
+
+  /**
+   * The human-readable name of the action plugin.
+   *
+   * @ingroup plugin_translatable
+   *
+   * @var \Drupal\Core\Annotation\Translation
+   */
+  public $label;
+
+  /**
+   * The category under which the action should be listed in the UI.
+   *
+   * @var \Drupal\Core\Annotation\Translation
+   *
+   * @ingroup plugin_translatable
+   */
+  public $category;
+
+  /**
+   * Defines the used context of the action plugin.
+   *
+   * Array keys are the names of the contexts and values context definitions.
+   *
+   * @var \Drupal\Core\Annotation\ContextDefinition[]
+   */
+  public $context = [];
+
+  /**
+   * Defines the provided context of the action plugin.
+   *
+   * Array keys are the names of the contexts and values context definitions.
+   *
+   * @var \Drupal\Core\Annotation\ContextDefinition[]
+   */
+  public $provides = [];
+}
