+++ tests/src/Integration/Action/EntityCreateTest.php
@@ -11,7 +11,7 @@ use Drupal\Core\Entity\ContentEntityType;
 use Drupal\Tests\rules\Integration\RulesEntityIntegrationTestBase;
 
 /**
- * @coversDefaultClass \Drupal\rules\Plugin\Action\EntityCreate
+ * @coversDefaultClass \Drupal\rules\Plugin\RulesAction\EntityCreate
  * @group rules_actions
  */
 class EntityCreateTest extends RulesEntityIntegrationTestBase {
