+++ tests/src/Integration/Action/EntitySaveTest.php
@@ -10,7 +10,7 @@ namespace Drupal\Tests\rules\Integration\Action;
 use Drupal\Tests\rules\Integration\RulesEntityIntegrationTestBase;
 
 /**
- * @coversDefaultClass \Drupal\rules\Plugin\Action\EntitySave
+ * @coversDefaultClass \Drupal\rules\Plugin\RulesAction\EntitySave
  * @group rules_actions
  */
 class EntitySaveTest extends RulesEntityIntegrationTestBase {
