+++ tests/src/Integration/Action/PathAliasCreateTest.php
@@ -11,7 +11,7 @@ use Drupal\Core\Language\LanguageInterface;
 use Drupal\Tests\rules\Integration\RulesIntegrationTestBase;
 
 /**
- * @coversDefaultClass \Drupal\rules\Plugin\Action\PathAliasCreate
+ * @coversDefaultClass \Drupal\rules\Plugin\RulesAction\PathAliasCreate
  * @group rules_actions
  */
 class PathAliasCreateTest extends RulesIntegrationTestBase {
