+++ tests/src/Integration/Action/PathAliasDeleteByAliasTest.php
@@ -10,7 +10,7 @@ namespace Drupal\Tests\rules\Integration\Action;
 use Drupal\Tests\rules\Integration\RulesIntegrationTestBase;
 
 /**
- * @coversDefaultClass \Drupal\rules\Plugin\Action\PathAliasDeleteByAlias
+ * @coversDefaultClass \Drupal\rules\Plugin\RulesAction\PathAliasDeleteByAlias
  * @group rules_actions
  */
 class PathAliasDeleteByAliasTest extends RulesIntegrationTestBase {
