+++ tests/src/Integration/Action/UserUnblockTest.php
@@ -10,7 +10,7 @@ namespace Drupal\Tests\rules\Integration\Action;
 use Drupal\Tests\rules\Integration\RulesEntityIntegrationTestBase;
 
 /**
- * @coversDefaultClass \Drupal\rules\Plugin\Action\UserUnblock
+ * @coversDefaultClass \Drupal\rules\Plugin\RulesAction\UserUnblock
  * @group rules_actions
  */
 class UserUnblockTest extends RulesEntityIntegrationTestBase {
