+++ tests/src/Integration/Action/VariableAddTest.php
@@ -10,7 +10,7 @@ namespace Drupal\Tests\rules\Integration\Action;
 use Drupal\Tests\rules\Integration\RulesIntegrationTestBase;
 
 /**
- * @coversDefaultClass \Drupal\rules\Plugin\Action\VariableAdd
+ * @coversDefaultClass \Drupal\rules\Plugin\RulesAction\VariableAdd
  * @group rules_actions
  */
 class VariableAddTest extends RulesIntegrationTestBase {
@@ -30,7 +30,7 @@ class VariableAddTest extends RulesIntegrationTestBase {
   public function testExecute() {
     $variable = $this->randomMachineName();
 
-    /** @var \Drupal\rules\Plugin\Action\VariableAdd $action */
+    /** @var \Drupal\rules\Plugin\RulesAction\VariableAdd $action */
     $action = $this->actionManager->createInstance('rules_variable_add');
     $action->setContextValue('value', $variable);
     $action->execute();
