+++ tests/src/Integration/RulesIntegrationTestBase.php
@@ -7,12 +7,12 @@
 
 namespace Drupal\Tests\rules\Integration;
 
-use Drupal\Core\Action\ActionManager;
 use Drupal\Core\Cache\NullBackend;
 use Drupal\Core\Condition\ConditionManager;
 use Drupal\Core\DependencyInjection\ContainerBuilder;
 use Drupal\Core\TypedData\TypedDataManager;
 use Drupal\rules\Context\DataProcessorManager;
+use Drupal\rules\Core\RulesActionManager;
 use Drupal\rules\Engine\ExpressionPluginManager;
 use Drupal\Tests\UnitTestCase;
 
@@ -37,7 +37,7 @@ abstract class RulesIntegrationTestBase extends UnitTestCase {
   protected $typedDataManager;
 
   /**
-   * @var \Drupal\Core\Action\ActionManager
+   * @var \Drupal\rules\Core\RulesActionManagerInterface
    */
   protected $actionManager;
 
@@ -127,7 +127,7 @@ abstract class RulesIntegrationTestBase extends UnitTestCase {
       'Drupal\\Core\\Validation' => $this->root . '/core/lib/Drupal/Core/Validation',
     ]);
 
-    $this->actionManager = new ActionManager($this->namespaces, $this->cacheBackend, $this->moduleHandler);
+    $this->actionManager = new RulesActionManager($this->namespaces, $this->cacheBackend, $this->moduleHandler);
     $this->conditionManager = new ConditionManager($this->namespaces, $this->cacheBackend, $this->moduleHandler);
     $this->rulesExpressionManager = new ExpressionPluginManager($this->namespaces, $this->moduleHandler);
 
@@ -151,7 +151,7 @@ abstract class RulesIntegrationTestBase extends UnitTestCase {
 
     $container->set('entity.manager', $this->entityManager);
     $container->set('path.alias_manager', $this->aliasManager);
-    $container->set('plugin.manager.action', $this->actionManager);
+    $container->set('plugin.manager.rules_action', $this->actionManager);
     $container->set('plugin.manager.condition', $this->conditionManager);
     $container->set('plugin.manager.rules_expression', $this->rulesExpressionManager);
     $container->set('plugin.manager.rules_data_processor', $this->rulesDataProcessorManager);
